dspam backend: allow configuring ignoring virus emails
The following new options needs to be added to the configuration for
anything to happen.

   antispam_dspam_result_header = X-DSPAM-Result
   antispam_dspam_result_blacklist = Virus

result_blacklist is a semicolon separated list of "blacklisted"
classifications and mails with the header X-DSPAM-Result set to
any of the strings (case insensitive) are ignored during move.

I've tested it and it seems to work as expected.