Tidy up debian/rules a bit
Drop the $(CURDIR) stuff, we've already tested that's where we are.
Copy the manpage just before using dh_installman on it.  There is no need to
move the original about (and possibly leave the source in an inconsistent
state if something fails in the wrong place).
Drop some calls to dh_* things that are a no-op here.
Don't install the upstream README, it's wrong for Debian users, and we already
have README.Debian which is more correct, if just as brief about it :)
Likewise the upstream NOTES aren't presently very interesting to users of the
binary package, leave it in the source.